fix: move icon styles completely to CSS
The X and O icons have default colors and sizes, but this approach has
some problems.

Sizes are in pixels, so they are not resized according to user
preferences. Furthermore, the design system colors are all in the
Tailwind CSS theme, so changing them there should change the colors of
the entire application, but this does not happen with the icons,
as the colors are hard-coded in their components.

This commit solves these issues by moving all icon styling to CSS, so
sizes and colors are now specified via classes, making icons more
accessible and the Tailwind CSS theme a single source of truth for the
design system.