fix: return to using only inner shadows
The design provided by Frontend Mentor only uses inner shadows, I
thought that shadows like this would make development difficult, but
I noticed that not using them made it harder to make the site look
like the design.

This commit solves this problem by changing the shadows in the
Tailwind CSS theme and some parts of the application, this way the
site only uses inner shadows and it is easier to make it look as
beautiful as the design.